// src/shared/enums/sla-categories.enum.ts
export enum SlaCategories {
  BUSINESS_HOURS = 'business_hours',
  AFTER_HOURS = 'after_hours',
  CRITICAL = 'critical',
  STANDARD = 'standard',
  LOW_PRIORITY = 'low_priority',
}

export enum TicketPriority {
  HIGH = 'high',
  MEDIUM = 'medium',
  LOW = 'low',
  CRITICAL = 'critical',
}

export enum TicketStatus {
  OPEN = 'open',
  IN_PROGRESS = 'in_progress',
  PENDING = 'pending',
  RESOLVED = 'resolved',
  CLOSED = 'closed',
}